+++ src/ui/drawing_view.py
@@ -1,7 +1,8 @@
-from tkinter import Button, Label, Canvas, Radiobutton, IntVar, font, DISABLED, NORMAL
+from tkinter import Button, Label, Canvas, Radiobutton, IntVar, DISABLED, NORMAL
 from ui.common import View
 from ui.prompt_text import PromptText
-from backend.backend import backend, RECTANGLE, OVAL, LINE, TEXT
+from backend.dwg_mgmt import dwg_mgr, RECTANGLE, OVAL, LINE, TEXT
+
 
 
 class DrawingView(View):
@@ -11,8 +12,7 @@ class DrawingView(View):
         """creates the main drawing view"""
 
         super().__init__(master, None, menu_view)
-        self._curr_user = backend.get_curr_user()
-        self._curr_dwg = backend.get_curr_dwg()
+        self._curr_dwg = dwg_mgr.get_curr_dwg()
         self._rows = 5
         self._create_widgets()
 
@@ -20,12 +20,12 @@ class DrawingView(View):
         """adds a button for border and a button for fill"""
 
         Button(self._frame, bg=color,
-               command=lambda: backend.set_border(color),
+               command=lambda: dwg_mgr.set_border(color),
                activebackground=color,
                ).grid(column=0, row=self._rows)
 
         Button(self._frame,  bg=color,
-               command=lambda: backend.set_fill(color),
+               command=lambda: dwg_mgr.set_fill(color),
                activebackground=color,
                ).grid(column=1, row=self._rows)
 
@@ -50,22 +50,22 @@ class DrawingView(View):
 
         Radiobutton(self._frame, text='rect',
                     value=RECTANGLE, variable=self._rb_val,
-                    command=lambda: backend.set_cmd(RECTANGLE)
+                    command=lambda: dwg_mgr.set_cmd(RECTANGLE)
                     ).grid(column=0, row=2)
 
         Radiobutton(self._frame, text='oval',
                     value=OVAL, variable=self._rb_val,
-                    command=lambda: backend.set_cmd(OVAL)
+                    command=lambda: dwg_mgr.set_cmd(OVAL)
                     ).grid(column=1, row=2)
 
         Radiobutton(self._frame, text='text',
                     value=TEXT, variable=self._rb_val,
-                    command=lambda: backend.set_cmd(TEXT)
+                    command=lambda: dwg_mgr.set_cmd(TEXT)
                     ).grid(column=0, row=3)
 
         Radiobutton(self._frame, text='line',
                     value=LINE, variable=self._rb_val,
-                    command=lambda: backend.set_cmd(LINE)
+                    command=lambda: dwg_mgr.set_cmd(LINE)
                     ).grid(column=1, row=3)
 
         Label(self._frame, text='↓ border - fill ↓').grid(columnspan=2, row=4)
@@ -83,53 +83,56 @@ class DrawingView(View):
             bg='white'
         )
         canv.grid(column=2, row=0, rowspan=20)
-        backend.set_cmd(RECTANGLE)
-        backend.set_border('red')
-        backend.set_fill('green')
+        dwg_mgr.set_cmd(RECTANGLE)
+        dwg_mgr.set_border('red')
+        dwg_mgr.set_fill('green')
 
         # separating app logic from UI here (?)
         self._master.title(f'Art + {self._curr_dwg.name}')
-        backend.set_canvas(canv, self.undo_btn_enabler)
-        canv.bind('<Button-1>', backend.b1_dn)
-        canv.bind('<B1-Motion>', backend.b1_mv)
-        canv.bind('<B1-ButtonRelease>', backend.b1_up)
+        dwg_mgr.set_canvas(canv, self.undo_btn_enabler)
+        canv.bind('<Button-1>', dwg_mgr.b1_dn)
+        canv.bind('<B1-Motion>', dwg_mgr.b1_mv)
+        canv.bind('<B1-ButtonRelease>', dwg_mgr.b1_up)
 
-        backend.set_text_prompter(self.prompt_text)
+        dwg_mgr.set_text_prompter(self.prompt_text)
 
         # add extra controls/functionalities
-        if self._curr_user.teacher:
-            pass
+        #if self._curr_user.teacher:
+        #    pass
 
     def _save_and_exit(self):
         """Saves the current drawing to database"""
 
         self._master.title('Art +')
-        backend.save_curr_dwg()
+        dwg_mgr.save_curr_dwg()
         self._handle_prev()
 
     def prompt_text(self):
         """
-            makes backend able to create a pop-up window
+            makes dwg_mgr able to create a pop-up window
             and prompt the user for text input
         """
 
-        prompt = PromptText(self._frame)
-        return prompt.process()
+        return PromptText(self._frame,
+            'Text to add',
+            {'text': 'text:'},
+            {},
+            {'text': 'Add text'}).get()
 
     def _undo(self):
         """pushes the last feature from the drawing to the undo stack"""
-        if backend.undo():
+        if dwg_mgr.undo():
             self._redo_btn['state'] = NORMAL
         else:
             self._undo_btn['state'] = DISABLED
 
     def _redo(self):
         """pushes 1 feature from undo stack to the drawing"""
-        if backend.redo():
+        if dwg_mgr.redo():
             self._undo_btn['state'] = NORMAL
         else:
             self._redo_btn['state'] = DISABLED
 
     def undo_btn_enabler(self):
-        """Makes Backend able to enable the undo button"""
+        """Makes dwg_mgr able to enable the undo button"""
         self._undo_btn['state'] = NORMAL