+++ src/ui/login_view.py
@@ -2,8 +2,8 @@ import re
 from textwrap import dedent
 from tkinter import messagebox, Entry, Button, Label, W
 from ui.common import View
-from ui.prompt_pw import PromptPassword
-from backend.backend import backend, WrongPassword
+from ui.prompt_text import PromptText
+from backend.user_mgmt import user_mgr, WrongPassword
 
 
 class LoginView(View):
@@ -63,12 +63,17 @@ class LoginView(View):
             return
 
         try:
-            backend.login_register(username, password, self.pw_confirmation)
+            user_mgr.login_register(username, password, self.pw_confirmation)
             self._handle_next()
 
         except WrongPassword:
             messagebox.showerror("Login/register failed", "wrong password")
 
     def pw_confirmation(self):
-        conf = PromptPassword(self._master, self._user.get())
-        return conf.process()
+        return PromptText(self._master,
+            f'Registering {self._user.get()}',
+            {'text': 'repeat password: '},
+            {'show': '*'},
+            {'text': 'Confirm user creation'}
+            ).get()
+