+++ src/ui/prompt_text.py
@@ -4,24 +4,24 @@ from tkinter import Toplevel, StringVar, Label, Entry, Button
 class PromptText(Toplevel):
     """basic input box for random text inputs"""
 
-    def __init__(self, master):
-        """"""
+    def __init__(self, master, title: str, kwlabel: dict, kwentry:dict, kwbutton:dict):
+
 
         super().__init__(master)
-        self.title('Text to input')
+        self.title(title)
 
-        self._text = StringVar(self, value='jotain')
-        Label(self, text='name:').grid(column=0, row=0)
-        txt = Entry(self, textvariable=self._text)
-        txt.bind('<Return>', lambda _ev: self.destroy())
-        txt.grid(column=1, row=0)
-        txt.focus_set()
+        self.__retval = StringVar(self)
+        Label(self, **kwlabel).grid(column=0, row=0)
+        entry = Entry(self, **kwentry, textvariable=self.__retval)
+        entry.bind('<Return>', lambda _ev: self.destroy())
+        entry.grid(column=1, row=0)
+        entry.focus_set()
 
-        Button(self, text='OK', command=self.destroy
+        Button(self, **kwbutton, command=self.destroy
                ).grid(columnspan=2, row=3)
 
-    def process(self):
-        """method for querying text input"""
-        # backend.set_next_text(self._text.get())
+    def get(self) -> str:
+        """Waits for the user input and returns it"""
+
         self.wait_window()
-        return self._text.get()
+        return self.__retval.get()